test: CdsApi moved to BuildSearchResult and HTTPDownload
