+++ tests/units/test_download_plugins.py
@@ -36,6 +36,7 @@ from tests import TEST_RESOURCES_PATH
 from tests.context import (
     DEFAULT_STREAM_REQUESTS_TIMEOUT,
     HTTP_REQ_TIMEOUT,
+    NOT_AVAILABLE,
     OFFLINE_STATUS,
     ONLINE_STATUS,
     USER_AGENT,
@@ -1006,6 +1007,88 @@ class TestDownloadPluginHttp(BaseDownloadPluginTest):
         )
         # empty product download directory should have been removed
         self.assertFalse(Path(os.path.join(self.output_dir, "dummy_product")).exists())
+
+    def test_plugins_download_http_order_download_cds(
+        self,
+    ):
+        """HTTPDownload.download must order the product if needed"""
+
+        self.product.provider = "cop_cds"
+        self.product.product_type = "CAMS_EAC4"
+        product_dataset = "cams-global-reanalysis-eac4"
+
+        plugin = self.get_download_plugin(self.product)
+        auth = self.get_auth_plugin(self.product.provider)
+        auth.config.credentials = {"username": "john", "password": "doe"}
+        self.product.register_downloader(plugin, auth)
+
+        endpoint = plugin.config.base_uri
+        self.product.properties["orderLink"] = (
+            f"{endpoint}/resources/{product_dataset}" + '?{"foo": "bar"}'
+        )
+        self.product.properties["storageStatus"] = "OFFLINE"
+        self.product.location = self.product.remote_location = (
+            NOT_AVAILABLE + '?{"foo": "bar"}'
+        )
+
+        @responses.activate(registry=responses.registries.OrderedRegistry)
+        def run():
+            responses.add(
+                responses.POST,
+                f"{endpoint}/resources/{product_dataset}",
+                status=200,
+                content_type="application/octet-stream",
+                body=b'{"state": "queued", "request_id": "dummy_request_id"}',
+                auto_calculate_content_length=True,
+            )
+            responses.add(
+                responses.GET,
+                f"{endpoint}/tasks/dummy_request_id",
+                status=200,
+                content_type="application/octet-stream",
+                body=b'{"state": "running", "request_id": "dummy_request_id"}',
+                auto_calculate_content_length=True,
+            )
+            responses.add(
+                responses.GET,
+                f"{endpoint}/tasks/dummy_request_id",
+                status=200,
+                content_type="application/octet-stream",
+                body=(
+                    b'{"state": "completed", '
+                    b'"request_id": "dummy_request_id", '
+                    b'"location": "http://somewhere/download/dummy_request_id"}'
+                ),
+                auto_calculate_content_length=True,
+            )
+            responses.add(
+                responses.GET,
+                "http://somewhere/download/dummy_request_id",
+                status=200,
+                content_type="application/octet-stream",
+                adding_headers={"content-disposition": ""},
+                body=b"some content",
+                auto_calculate_content_length=True,
+            )
+
+            output_data_path = self.output_dir
+
+            # expected values
+            expected_remote_location = "http://somewhere/download/dummy_request_id"
+            expected_path = os.path.join(
+                output_data_path, self.product.properties["title"]
+            )
+            # download
+            path = self.product.download(
+                outputs_prefix=output_data_path,
+                wait=0.001 / 60,
+                timeout=0.2 / 60,
+            )
+            self.assertEqual(self.product.remote_location, expected_remote_location)
+            self.assertEqual(path, expected_path)
+            self.assertEqual(self.product.location, path_to_uri(expected_path))
+
+        run()
 
     @mock.patch("eodag.plugins.download.http.requests.request", autospec=True)
     def test_plugins_download_http_order_get(self, mock_request):
